+++ inbm/dispatcher-agent/tests/unit/test_ota_factory.py
@@ -1,3 +1,5 @@
+import threading
+
 import pytest
 
 from unit.common.mock_resources import *
@@ -28,7 +30,8 @@ def test_get_factory(ota_type, expected_factory, mock_disp_obj, mock_disp_broker
         None,
         MockInstallCheckService(),
         UpdateLogger(ota_type=ota_type, data="metadata"),
-        ConfigDbs.ON
+        ConfigDbs.ON,
+        cancel_event=threading.Event()
     )
     assert isinstance(factory, expected_factory)
 
@@ -43,7 +46,8 @@ def test_raise_error_unsupported_ota(mock_disp_obj, mock_disp_broker) -> None:
             None,
             MockInstallCheckService(),
             UpdateLogger(ota_type="IOTA", data="metadata"),
-            ConfigDbs.OFF
+            ConfigDbs.OFF,
+            cancel_event=threading.Event()
         )
 
 
@@ -61,7 +65,8 @@ def test_create_parser(ota_type, expected_parser, mock_disp_obj, mock_disp_broke
         None,
         MockInstallCheckService(),
         UpdateLogger(ota_type=ota_type, data="metadata"),
-        ConfigDbs.ON
+        ConfigDbs.ON,
+        cancel_event=threading.Event()
     ).create_parser()
     assert isinstance(parser, expected_parser)
 
@@ -80,6 +85,7 @@ def test_create_thread(ota_type, expected_thread, mock_disp_obj, mock_disp_broke
         None,
         MockInstallCheckService(),
         UpdateLogger(ota_type=ota_type, data="metadata"),
-        ConfigDbs.ON
+        ConfigDbs.ON,
+        cancel_event=threading.Event()
     ).create_thread({'abc': 'def'})
     assert isinstance(thread, expected_thread)