+++ inbm/Changelog.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## NEXT - YYYY-MM-DD
+### Changed
+ - (NEX-14264) Download file to /opt directory in TiberOS
 
 ## 4.2.6.1 - 2024-10-18
 ### Added