+++ inbm/Changelog.md
@@ -4,8 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## NEXT - YYYY-MM-DD
-### Added
+ - (NEXMANAGE-837) Fix schema validation issue on failed UDM request sent from
+ dispatcher to cloudadapter
  - (NEXMANAGE-900) Add UpdateFirmware to the common proto file
+
 
 ## 4.2.6.1 - 2024-10-18
 ### Added